Move LazySeriesIterator to chunk package
